# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_ademarthiago_payment-gateway_internal_usecase.GetPaymentOutput:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      external_id:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      provider:
        type: string
      status:
        type: string
      transactions:
        items:
          $ref: '#/definitions/github_com_ademarthiago_payment-gateway_internal_usecase.TransactionOutput'
        type: array
      updated_at:
        type: string
    type: object
  github_com_ademarthiago_payment-gateway_internal_usecase.ProcessRefundOutput:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      payment_id:
        type: string
      status:
        type: string
      transaction_id:
        type: string
    type: object
  github_com_ademarthiago_payment-gateway_internal_usecase.TransactionOutput:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      provider_ref:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  internal_adapter_http_handler.createPaymentRequest:
    properties:
      amount:
        type: integer
      currency:
        type: string
      description:
        type: string
      external_id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      provider:
        type: string
    type: object
  internal_adapter_http_handler.createPaymentResponse:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      external_id:
        type: string
      id:
        type: string
      provider:
        type: string
      status:
        type: string
    type: object
  internal_adapter_http_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  internal_adapter_http_handler.refundRequest:
    properties:
      amount:
        type: integer
      reason:
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  description: A payment gateway POC built with Go and Hexagonal Architecture
  title: Payment Gateway API
  version: "1.0"
paths:
  /payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_http_handler.createPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapter_http_handler.createPaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http_handler.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_adapter_http_handler.errorResponse'
      summary: Create a payment
      tags:
      - payments
  /payments/{id}:
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ademarthiago_payment-gateway_internal_usecase.GetPaymentOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http_handler.errorResponse'
      summary: Get a payment by ID
      tags:
      - payments
  /payments/{id}/refund:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Refund request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_http_handler.refundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ademarthiago_payment-gateway_internal_usecase.ProcessRefundOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http_handler.errorResponse'
      summary: Process a refund
      tags:
      - payments
swagger: "2.0"